Revise FINE-level request query logging algorithm
Came up with a better approach: if a large query document and operation name is
given, then use a quick-and-dirty regex search to locate the
query/mutation/subscription within it and print that out as an excerpt; this
allows the effective query to be shown in the logs, which saves me from having
to look up the source file it came from.

If the regex search fails (it's not meant to be perfect), then either print out
the entire query document verbatim (if less than 500 characters), otherwise
abbreviate it to a single line as was done before.

n.b. Single-operation queries without an operation name are still printed
in full, as before.